### 2021-11-18 #47 @enra64
Presubmission Inquiry: Visualization Tool: Sevivi, a python package and CLI tool to generate videos of sensor data graphs synchronized to a video of the sensor movement
Submitting Author: Arne Herdick (@enra64)
Package Name: sevivi
One-Line Description of Package: sevivi is a python package and command line tool to generate videos of sensor data graphs synchronized to a video of the sensor movement.
Repository Link (if existing):   https://github.com/hpi-ch/sevivi/

---

Sevivi is designed to render plots of sensor data next to a video that was taken synchronously, synchronizing the sensor data precisely to the video. It allows you to investigate why certain patterns occur in your sensor data based on the exact moment in the video.

Sevivi renders plots of given data next to a given video.

Target audience is researchers working with motion data, e.g., prediction of squat intensity using acceleration data. When these researchers have taken videos of their trials, they might want to see what exactly has produced a certain pattern, helping to differentiate between noise and signal. Sevivi makes this easier, by rendering synchronized videos of both the original video and the sensor data plots synchronously. Synchronization can be done manually, using an IMU on the camera (e.g. https://github.com/DavidGillsjo/VideoIMUCapture-Android/) or by using skeleton data from a tracking software (we tested with an azure kinect).

Our research indicates that no similar python packages or other programs exist.
