### 2021-11-23 #47 @NickleDave
Hi @enra64 welcome to pyOpenSci!

Thank you for your detailed pre-submission inquiry.

I need to check in with other editors about where we are at with software reviews, but I did want to make sure you knew we saw this.

It's about to be a holiday where I am in the US, but we will get back to by the end of next week.
