### 2021-12-16 #47 @NickleDave
Hi again @enra64 -- sorry for the delay in getting back to you.

This does appear to be within scope as a data visualization tool.

It's not clear to me who we would reach out to for reviews though.

Would you be able to point me to some researchers working in related areas, or developing related tools?
I realize you did not find any similar programs that exist for your specific tool, but if you can give me an idea of who the "target audience" that would really help. E.g., if you can think of 2 or 3 researchers in your area.

Thank you
