### 2021-12-24 #47 @lwasser
hi @edgarriba we are hitting the holiday season here in the US and so pyOpenSci will be taking a holiday break through the new year. Here is what i suggest before I sign off for the next week! First, please open a new submission for review if your package is ready.

Please do this anytime but know that we won't be able to look at it until after January 3.
Once you do that I will close this issue and we can begin looking for reviewers. Until then, please have a wonderful new year and we will check back in here after January 3.
I am signing off now so i won't see any additional messages until then as well.
