### 2021-12-24 #47 @NickleDave
Hi @edgarriba thank you for your quick reply.

In our guide we ask for a three week turnaround but with everyone's schedules extra taxed right now I would say 1-2 months.

I am actually waiting to hear more from the author of the package @enra64 about the intended audience -- I think that might have got lost in the conversation above.

Let me hear more from @enra64 and do my own homework based on related tools from @arianesasso then discuss with @lwasser after the holidays.

So, long story short, we are not asking you to review right now but if we did it, we'd hope to have it back 1-2 months after the initial submission.
