### 2021-12-27 #47 @enra64
Thanks everyone for checking this out. Hopefully my colleague will take over creating the review issue soon. Regarding the related researchers: I think the following papers from the exercise recognition/exercise intensity recognition space might be interesting:

* [Pernek et al.: Recognizing the Intensity of Strength Training Exercises with Wearable Sensors](https://www.researchgate.net/publication/282731812_Recognizing_the_Intensity_of_Strength_Training_Exercises_with_Wearable_Sensors)
* Our own paper [Using Machine Learning to Predict
Perceived Exertion During Resistance Training With Wearable Heart Rate and Movement Sensors](https://www13.hpi.uni-potsdam.de/fileadmin/user_upload/fachgebiete/arnrich/publications/2021_IEEE_BIBM_RPE_Prediction_IMU_and_HRV.pdf)
* [Jiang et al.: A data-
driven approach to predict fatigue in exercise based on motion data from
wearable sensors or force plate](https://www.mdpi.com/1424-8220/21/4/1499)

Another area of interest for this package might be gait analysis, even though this could be difficult due to moving subjects.
